requirements.txt
Flask==3.0.3
googletrans==4.0.0rc1
Pillow==10.4.0
protobuf==3.12.4
python-dotenv==1.0.1
Requests==2.32.3
retry==0.9.2
twilio==9.0.4
